requirements.txt
# Core Application
fastapi==0.95.2  # Slightly older but very stable
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Video Processing
ffmpeg-python==0.2.0
moviepy==1.0.3
opencv-python-headless==4.6.0.66

# AI/ML Stack (fully compatible set)
torch==1.13.1
numpy==1.23.5
tiktoken==0.3.3
openai-whisper==20231117  #20230308  # Stable version

# Supporting Libraries (compatible versions)
certifi==2022.12.7
charset-normalizer==3.0.1
filelock==3.9.0
fsspec==2023.3.0
future==1.0.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.2.1
networkx==2.8.8
regex==2022.10.31
requests==2.28.2
sympy==1.11.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15

